feat: add data loader to seed database on startup

+++ backend/src/main/java/com/boomerang/admindashboard/dto/CustomerDTO.java
@@ -2,6 +2,7 @@ package com.boomerang.admindashboard.dto;
 
 import lombok.Data;
 import java.time.LocalDate;
+import java.math.BigDecimal;
 
 @Data
 public class CustomerDTO {
@@ -9,5 +10,5 @@ public class CustomerDTO {
     private String fullName;
     private String email;
     private LocalDate joinDate;
-    private Double totalSpend;
+    private BigDecimal totalSpend;
 }
