+++ backend/src/main/java/com/boomerang/admindashboard/dto/DashboardStatsDTO.java
@@ -2,11 +2,12 @@ package com.boomerang.admindashboard.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class DashboardStatsDTO {
-    private double totalSales;
+    private BigDecimal totalSales;
     private long totalBookings; // Assuming total bookings == total sales records for now
     private long newCustomersThisMonth;
 }
